+++ src/components/Button/index.tsx
@@ -1,25 +1,8 @@
-import {
-  AnchorHTMLAttributes,
-  ButtonHTMLAttributes,
-  ElementType,
-  ForwardRefRenderFunction,
-  ReactNode,
-  forwardRef
-} from "react";
+import { ForwardRefRenderFunction, forwardRef } from "react";
+
 import * as S from "./styles";
 
-type ButtonType =
-  | ButtonHTMLAttributes<HTMLButtonElement>
-  | AnchorHTMLAttributes<HTMLAnchorElement>;
-
-export type ButtonProps = {
-  variant?: "primary" | "neutral" | "ghost" | "danger";
-  size?: "small" | "medium" | "full";
-  children: ReactNode;
-  as?: ElementType;
-} & ButtonType;
-
-export type ButtonContentProps = Pick<ButtonProps, "size" | "variant">;
+import { ButtonContentProps, ButtonProps } from "./types";
 
 const Button: ForwardRefRenderFunction<ButtonContentProps, ButtonProps> = (
   { children, size = "medium", variant = "primary", ...props },
