+++ src/components/Button/types.d.ts
@@ -1,0 +1,19 @@
+import {
+  AnchorHTMLAttributes,
+  ButtonHTMLAttributes,
+  ElementType,
+  ReactNode
+} from "react";
+
+export type ButtonType =
+  | ButtonHTMLAttributes<HTMLButtonElement>
+  | AnchorHTMLAttributes<HTMLAnchorElement>;
+
+export type ButtonProps = {
+  variant?: "primary" | "neutral" | "ghost" | "danger";
+  size?: "small" | "medium" | "full";
+  children: ReactNode;
+  as?: ElementType;
+} & ButtonType;
+
+export type ButtonContentProps = Pick<ButtonProps, "size" | "variant">;
